codes/django-study/requirements.txt
Django
djangorestframework
psycopg2-binary

pytest
pytest-django

ipython
